Fix #1
Our plugin saves the custom order number as a post meta for the order.
When a renewal order is created by WC Subscriptions, the meta fields are
copied. The order number also gets copied and hence the issue.

This has now been fixed by ensuring the custom order number is updated
when a renewal order is created.